### 2025-06-06 #8367 @editorialbot
[PRE REVIEW]: kokkos-fft: A shared-memory FFT for the Kokkos ecosystem
**Submitting author:** @yasahi-hpc (<a href="http://orcid.org/0000-0002-9997-1274">Yuuichi ASAHI</a>)
**Repository:** https://github.com/kokkos/kokkos-fft
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.0
