### 2025-06-10 11ffc36
rename BinaryOp and fix docstring issues
