+++ testing/src/KokkosFFT_CountErrors.hpp
@@ -48,11 +48,12 @@ Kokkos::Iterate get_iteration_order(const ViewType& view) {
 /// parallel_reduce.
 /// \tparam AViewType The type of the first Kokkos view.
 /// \tparam BViewType The type of the second Kokkos view.
-/// \tparam BinaryOp The binary operation to apply to the elements of the views.
+/// \tparam ComparisonOp The binary operation to apply to the elements of the
+/// views.
 /// \tparam Layout The layout type of the Kokkos views.
 /// \tparam iType The integer type used for indexing the view elements.
 template <KokkosExecutionSpace ExecutionSpace, KokkosView AViewType,
-          KokkosView BViewType, typename BinaryOp, KokkosLayout Layout,
+          KokkosView BViewType, typename ComparisonOp, KokkosLayout Layout,
           typename iType>
 struct ViewErrors {
  private:
@@ -104,14 +105,14 @@ struct ViewErrors {
   ///
   /// \param[in] a First Kokkos view containing data to compare.
   /// \param[in] b Second Kokkos view containing data to compare against.
-  /// \param[in] p The binary operation used for comparison
+  /// \param[in] op The binary operation used for comparison
   /// \param[in] space The Kokkos execution space used to launch the parallel
   /// reduction.
-  ViewErrors(const AViewType& a, const BViewType& b, BinaryOp p,
+  ViewErrors(const AViewType& a, const BViewType& b, ComparisonOp op,
              const ExecutionSpace space = ExecutionSpace()) {
     Kokkos::parallel_reduce(
         "ViewErrors", get_policy(space, a),
-        CountErrors<std::make_index_sequence<m_rank_truncated>>(a, b, p),
+        CountErrors<std::make_index_sequence<m_rank_truncated>>(a, b, op),
         m_error);
   }
 
@@ -129,15 +130,15 @@ struct ViewErrors {
     AViewType m_a;
     BViewType m_b;
 
-    BinaryOp m_p;
+    ComparisonOp m_op;
 
     /// \brief Constructs the error counter and performs the error computation.
     ///
     /// \param[in] a First Kokkos view containing data to compare.
     /// \param[in] b Second Kokkos view containing data to compare against.
     /// \param[in] p The binary operation used for comparison
-    CountErrors(const AViewType& a, const BViewType& b, BinaryOp p)
-        : m_a(a), m_b(b), m_p(p) {}
+    CountErrors(const AViewType& a, const BViewType& b, ComparisonOp op)
+        : m_a(a), m_b(b), m_op(op) {}
 
     /// \brief Operator called by Kokkos to perform the comparison of each
     /// element.
@@ -148,13 +149,13 @@ struct ViewErrors {
     KOKKOS_INLINE_FUNCTION
     void operator()(const IndicesType<Idx>... indices, std::size_t& err) const {
       if constexpr (AViewType::rank() <= 6) {
-        bool close = m_p(m_a(indices...), m_b(indices...));
+        bool close = m_op(m_a(indices...), m_b(indices...));
         err += static_cast<std::size_t>(!close);
       } else if constexpr (AViewType::rank() == 7) {
         for (iType i6 = 0; i6 < iType(m_a.extent(6)); i6++) {
           auto tmp_a = m_a(indices..., i6);
           auto tmp_b = m_b(indices..., i6);
-          bool close = m_p(tmp_a, tmp_b);
+          bool close = m_op(tmp_a, tmp_b);
           err += static_cast<std::size_t>(!close);
         }
       } else if constexpr (AViewType::rank() == 8) {
@@ -162,7 +163,7 @@ struct ViewErrors {
[…]
