+++ testing/src/KokkosFFT_Ulps.hpp
@@ -114,12 +114,12 @@ KOKKOS_INLINE_FUNCTION bool almost_equal_ulps(ScalarA a, ScalarB b,
 }
 
 template <typename IntType>
-struct UlpsComparisonOP {
+struct UlpsComparisonOp {
  private:
   IntType m_max_ulps_diff;
 
  public:
-  UlpsComparisonOP(IntType max_ulps_diff = 1)
+  UlpsComparisonOp(IntType max_ulps_diff = 1)
       : m_max_ulps_diff(max_ulps_diff) {}
 
   template <typename ScalarA, typename ScalarB>
