### 2025-06-10 bda0214
Generalize count and find errors to work on a comparison operator (#279)
* generalize count and find errors to work on a comparison operator

* Rename AlmostEqualUlpsOp to UlpsComparisonOP

* rename BinaryOp and fix docstring issues

* rename UlpsComparisonOP into UlpsComparisonOp

* remove unused headers in KokkosFFT_AlmostEqual.hpp
