+++ testing/src/KokkosFFT_CountErrors.hpp
@@ -13,12 +13,6 @@ namespace KokkosFFT {
 namespace Testing {
 namespace Impl {
 
-template <typename ScalarA, typename ScalarB, typename ScalarTol>
-KOKKOS_INLINE_FUNCTION bool are_not_close(ScalarA a, ScalarB b, ScalarTol rtol,
-                                          ScalarTol atol) {
-  return Kokkos::abs(a - b) > (atol + rtol * Kokkos::abs(b));
-}
-
 template <KokkosView ViewType>
 Kokkos::Iterate get_iteration_order(const ViewType& view) {
   int64_t strides[ViewType::rank + 1];
@@ -54,10 +48,13 @@ Kokkos::Iterate get_iteration_order(const ViewType& view) {
 /// parallel_reduce.
 /// \tparam AViewType The type of the first Kokkos view.
 /// \tparam BViewType The type of the second Kokkos view.
+/// \tparam ComparisonOp The binary operation to apply to the elements of the
+/// views.
 /// \tparam Layout The layout type of the Kokkos views.
 /// \tparam iType The integer type used for indexing the view elements.
 template <KokkosExecutionSpace ExecutionSpace, KokkosView AViewType,
-          KokkosView BViewType, KokkosLayout Layout, typename iType>
+          KokkosView BViewType, typename ComparisonOp, KokkosLayout Layout,
+          typename iType>
 struct ViewErrors {
  private:
   // Since MDRangePolicy is not available for 7D and 8D views, we need to
@@ -108,19 +105,14 @@ struct ViewErrors {
   ///
   /// \param[in] a First Kokkos view containing data to compare.
   /// \param[in] b Second Kokkos view containing data to compare against.
-  /// \param[in] rtol Relative tolerance for comparing the view elements
-  /// (default 1.e-5).
-  /// \param[in] atol Absolute tolerance for comparing the view elements
-  /// (default 1.e-8).
+  /// \param[in] op The binary operation used for comparison
   /// \param[in] space The Kokkos execution space used to launch the parallel
   /// reduction.
-  ViewErrors(const AViewType& a, const BViewType& b, double rtol = 1.e-5,
-             double atol                = 1.e-8,
+  ViewErrors(const AViewType& a, const BViewType& b, ComparisonOp op,
              const ExecutionSpace space = ExecutionSpace()) {
     Kokkos::parallel_reduce(
         "ViewErrors", get_policy(space, a),
-        CountErrors<std::make_index_sequence<m_rank_truncated>>(a, b, rtol,
-                                                                atol),
+        CountErrors<std::make_index_sequence<m_rank_truncated>>(a, b, op),
         m_error);
   }
 
@@ -138,20 +130,15 @@ struct ViewErrors {
     AViewType m_a;
     BViewType m_b;
 
-    double m_rtol;
-    double m_atol;
+    ComparisonOp m_op;
 
     /// \brief Constructs the error counter and performs the error computation.
     ///
     /// \param[in] a First Kokkos view containing data to compare.
     /// \param[in] b Second Kokkos view containing data to compare against.
-    /// \param[in] rtol Relative tolerance for comparing the view elements
-    /// (default 1.e-5).
-    /// \param[in] atol Absolute tolerance for comparing the view elements
-    /// (default 1.e-8).
-    CountErrors(const AViewType& a, const BViewType& b, double rtol = 1.e-5,
-                double atol = 1.e-8)
-        : m_a(a), m_b(b), m_rtol(rtol), m_atol(atol) {}
+    /// \param[in] p The binary operation used for comparison
+    CountErrors(const AViewType& a, const BViewType& b, ComparisonOp op)
+        : m_a(a), m_b(b), m_op(op) {}
 
[…]
