### 2025-06-10 #8367 @danielskatz
👋  @thivinanandh & @jpmorgan98 & @jdoerfert, would one or more of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html (under our [COI](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and [CoC](https://joss.theoj.org/about#code_of_conduct) policies)
