### 2025-06-11 #8367 @jpmorgan98
@danielskatz I am available for review starting at the beginning of July if that timeline is ok
