### 2025-06-18 ffc58f6
use Kokkos::fabs to deal with half types
