### 2025-06-18 5d4b774
Implement AlmostEqualNulp and corresponding tests
