+++ testing/unit_test/Test_Utils.hpp
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: (C) The kokkos-fft development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT OR Apache-2.0 WITH LLVM-exception
+
 #ifndef TEST_UTILS_HPP
 #define TEST_UTILS_HPP
 
