### 2025-06-19 fac3f25
rename verbose to max_displayed_errors
