### 2025-06-19 5d9a8c1
fix: docstrings
