### 2025-06-19 8ccad3a
Almost equal nulp (#285)
* use Kokkos::fabs to deal with half types

* move fp16 wrapper into Test_Utils.hpp

* Implement AlmostEqualNulp and corresponding tests

* Add license in Test_Utils.hpp

* rename verbose to max_displayed_errors

* fix: docstrings
