+++ testing/src/KokkosFFT_AlmostEqualNulp.hpp
@@ -2,14 +2,14 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0 WITH LLVM-exception
 
-#ifndef KOKKOSFFT_ALLCLOSE_HPP
-#define KOKKOSFFT_ALLCLOSE_HPP
+#ifndef KOKKOSFFT_ALMOST_EQUAL_NULP_HPP
+#define KOKKOSFFT_ALMOST_EQUAL_NULP_HPP
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 #include <Kokkos_Core.hpp>
 #include "KokkosFFT_Concepts.hpp"
-#include "KokkosFFT_AlmostEqual.hpp"
+#include "KokkosFFT_Ulps.hpp"
 #include "KokkosFFT_CountErrors.hpp"
 #include "KokkosFFT_PrintErrors.hpp"
 
@@ -17,7 +17,7 @@ namespace KokkosFFT {
 namespace Testing {
 namespace Impl {
 /// \brief Compares two Kokkos views element-wise and checks if they are close
-///        within specified relative and absolute tolerances.
+///        within specified n units in the last place (ULPs).
 ///        AViewType and BViewType must have the same rank,
 ///        non_const_value_type, and execution_space;
 ///
@@ -27,10 +27,8 @@ namespace Impl {
 /// \param[out] listener The testing match result listener.
 /// \param[in] actual The actual Kokkos view.
 /// \param[in] expected The expected (reference) Kokkos view.
-/// \param[in] rtol Relative tolerance for comparing the view elements
-/// (default 1.e-5).
-/// \param[in] atol Absolute tolerance for comparing the view elements
-/// (default 1.e-8).
+/// \param[in] nulp The maximum allowed difference in ULPs for the
+/// numbers to be considered equal
 /// \param[in] max_displayed_errors How many elements to be reported
 /// (default: 3)
 template <KokkosView AViewType, KokkosView BViewType>
@@ -39,10 +37,10 @@ template <KokkosView AViewType, KokkosView BViewType>
            std::is_same_v<typename AViewType::non_const_value_type,
                           typename BViewType::non_const_value_type> &&
            (AViewType::rank() == BViewType::rank()))
-inline bool allclose_impl(testing::MatchResultListener* listener,
-                          const AViewType& actual, const BViewType& expected,
-                          double rtol, double atol,
-                          std::size_t max_displayed_errors) {
+inline bool almost_equal_nulp_impl(testing::MatchResultListener* listener,
+                                   const AViewType& actual,
+                                   const BViewType& expected, std::size_t nulp,
+                                   std::size_t max_displayed_errors) {
   const std::size_t rank = actual.rank();
   for (std::size_t i = 0; i < rank; i++) {
     if (actual.extent(i) != expected.extent(i)) {
@@ -54,7 +52,7 @@ inline bool allclose_impl(testing::MatchResultListener* listener,
   using ExecutionSpace = typename AViewType::execution_space;
   ExecutionSpace exec_space;
 
-  KokkosFFT::Testing::Impl::AlmostEqualOp op(rtol, atol);
+  KokkosFFT::Testing::Impl::UlpsComparisonOp op(nulp);
 
   std::size_t errors =
       KokkosFFT::Testing::Impl::count_errors(exec_space, actual, expected, op);
@@ -73,21 +71,17 @@ inline bool allclose_impl(testing::MatchResultListener* listener,
 }
 }  // namespace Impl
 
-MATCHER_P4(allclose, expected, rtol, atol, max_displayed_errors, "") {
[…]
