### 2025-06-25 9b87ca9
Refactor convert_negative_axis to unuse View
