### 2025-06-25 6e5e244
Implicitly cast std::size_t to int
