### 2025-06-25 32b9a14
Refactor convert_negative_axis to unuse View (#286)
* Refactor convert_negative_axis to unuse View

* We do not use Layout in the ConvertNegativeAxis tests

* Implicitly cast std::size_t to int

---------

Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
