+++ common/src/KokkosFFT_padding.hpp
@@ -39,13 +39,13 @@ auto get_modified_shape(const InViewType in, const OutViewType /* out */,
   }
 
   // Convert the input axes to be in the range of [0, rank-1]
+  constexpr int rank = static_cast<int>(InViewType::rank());
   std::vector<int> positive_axes;
   for (std::size_t i = 0; i < DIM; i++) {
-    int axis = KokkosFFT::Impl::convert_negative_axis(in, axes.at(i));
+    int axis = KokkosFFT::Impl::convert_negative_axis(rank, axes.at(i));
     positive_axes.push_back(axis);
   }
 
-  constexpr int rank    = static_cast<int>(InViewType::rank());
   using full_shape_type = shape_type<rank>;
   full_shape_type modified_shape;
   for (int i = 0; i < rank; i++) {
