+++ common/unit_test/Test_Common_Utils.cpp
@@ -21,11 +21,6 @@ using paired_scalar_types = ::testing::Types<
 
 // Basically the same fixtures, used for labeling tests
 template <typename T>
-struct ConvertNegativeAxis : public ::testing::Test {
-  using layout_type = T;
-};
-
-template <typename T>
 struct ConvertNegativeShift : public ::testing::Test {
   using layout_type = T;
 };
@@ -45,15 +40,12 @@ struct PairedScalarTypes : public ::testing::Test {
 };
 
 // Tests for convert_negative_axes over ND views
-template <typename LayoutType>
 void test_convert_negative_axes_1d() {
-  const int len        = 30;
-  using RealView1Dtype = Kokkos::View<double*, LayoutType, execution_space>;
-  RealView1Dtype x("x", len);
-
-  int converted_axis_0 = KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/0);
+  const std::size_t DIM = 1;
+  int converted_axis_0 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/0);
   int converted_axis_minus1 =
-      KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/-1);
+      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/-1);
 
   int ref_converted_axis_0      = 0;
   int ref_converted_axis_minus1 = 0;
@@ -63,23 +55,21 @@ void test_convert_negative_axes_1d() {
 
   // Check if errors are correctly raised against invalid axis
   // axis must be in [-1, 1)
-  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/1); },
+  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/1); },
                std::runtime_error);
 
-  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/-2); },
+  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/-2); },
                std::runtime_error);
 }
 
-template <typename LayoutType>
 void test_convert_negative_axes_2d() {
-  const int n0 = 3, n1 = 5;
-  using RealView2Dtype = Kokkos::View<double**, LayoutType, execution_space>;
-  RealView2Dtype x("x", n0, n1);
-
-  int converted_axis_0 = KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/0);
-  int converted_axis_1 = KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/1);
+  const std::size_t DIM = 2;
+  int converted_axis_0 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/0);
+  int converted_axis_1 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/1);
   int converted_axis_minus1 =
-      KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/-1);
+      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/-1);
 
   int ref_converted_axis_0      = 0;
   int ref_converted_axis_1      = 1;
@@ -91,26 +81,25 @@ void test_convert_negative_axes_2d() {
 
   // Check if errors are correctly raised against invalid axis
   // axis must be in [-2, 2)
-  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/2); },
[…]
