### 2025-06-25 083cdbd
Allow index_sequence to work on unsigned integers
