### 2025-06-25 08bf627
rename values in test_index_sequence test
