### 2025-06-26 801b235
remove unused views from index_sequence test
