### 2025-06-26 95a4121
make convert_negative_axis a templated function again
