+++ common/unit_test/Test_Common_Utils.cpp
@@ -10,7 +10,8 @@ using execution_space = Kokkos::DefaultExecutionSpace;
 using test_types = ::testing::Types<Kokkos::LayoutLeft, Kokkos::LayoutRight>;
 
 // Int like types
-using base_int_types = ::testing::Types<int, std::size_t>;
+using base_int_types   = ::testing::Types<int, std::size_t>;
+using signed_int_types = ::testing::Types<int, std::int32_t, std::int64_t>;
 
 // value type combinations that are tested
 using paired_scalar_types = ::testing::Types<
@@ -20,6 +21,11 @@ using paired_scalar_types = ::testing::Types<
     std::pair<Kokkos::complex<double>, Kokkos::complex<double>>>;
 
 // Basically the same fixtures, used for labeling tests
+template <typename T>
+struct TestConvertNegativeAxis : public ::testing::Test {
+  using value_type = T;
+};
+
 template <typename T>
 struct ConvertNegativeShift : public ::testing::Test {
   using layout_type = T;
@@ -40,40 +46,42 @@ struct PairedScalarTypes : public ::testing::Test {
 };
 
 // Tests for convert_negative_axes over ND views
+template <typename IntType>
 void test_convert_negative_axes_1d() {
-  const std::size_t DIM = 1;
-  int converted_axis_0 =
-      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/0);
-  int converted_axis_minus1 =
-      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/-1);
+  const IntType DIM = 1;
+  IntType converted_axis_0 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, IntType(0));
+  IntType converted_axis_minus1 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, IntType(-1));
 
-  int ref_converted_axis_0      = 0;
-  int ref_converted_axis_minus1 = 0;
+  IntType ref_converted_axis_0      = 0;
+  IntType ref_converted_axis_minus1 = 0;
 
   EXPECT_EQ(converted_axis_0, ref_converted_axis_0);
   EXPECT_EQ(converted_axis_minus1, ref_converted_axis_minus1);
 
   // Check if errors are correctly raised against invalid axis
   // axis must be in [-1, 1)
-  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/1); },
+  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, IntType(1)); },
                std::runtime_error);
 
-  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/-2); },
+  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, IntType(-2)); },
                std::runtime_error);
 }
 
+template <typename IntType>
 void test_convert_negative_axes_2d() {
-  const std::size_t DIM = 2;
-  int converted_axis_0 =
-      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/0);
-  int converted_axis_1 =
-      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/1);
-  int converted_axis_minus1 =
-      KokkosFFT::Impl::convert_negative_axis(DIM, /*axis=*/-1);
+  const IntType DIM = 2;
+  IntType converted_axis_0 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, IntType(0));
[…]
