### 2025-06-26 0a05d46
Make View rank a template parameter
