### 2025-06-26 ab0f746
Revert "Refactor convert_negative_axis to unuse View (#286)"
This reverts commit 32b9a14592ff7c5c9b56a8feeeb25decd0b92ffe.
