### 2025-06-27 4079e9e
Include config.h in KokkosFFT_default_types.hpp
