+++ CMakeLists.txt
@@ -95,6 +95,12 @@ if(KOKKOSFFT_TPL_LIST)
     # [TO DO] show more information about the library (like location)
     pad_string(BACKEND_PADDED "${BACKEND}" 13)
     message(STATUS "    ${BACKEND_PADDED}: ${TPL}")
+
+    # Rename FFTW_<BACKEND> -> FFTW
+    string(REGEX REPLACE "_.*" "" TPL_NAME "${TPL}")
+
+    # Here we set the global variable
+    set(KOKKOSFFT_ENABLE_TPL_${TPL_NAME} TRUE)
   endforeach()
 else()
   message(STATUS " (None)")
@@ -138,7 +144,10 @@ endif()
 set(PACKAGE_NAME_CONFIG_FILES KokkosFFT_config.h KokkosFFT_Version_Info.hpp)
 
 foreach(CONFIG_FILE ${PACKAGE_NAME_CONFIG_FILES})
-  configure_file(cmake/${CONFIG_FILE}.in ${CMAKE_BINARY_DIR}/${CONFIG_FILE})
+  configure_file(cmake/${CONFIG_FILE}.in
+                 ${CMAKE_CURRENT_BINARY_DIR}/fft/src/${CONFIG_FILE}
+                 @ONLY
+  )
 endforeach()
 
 # Set directories used for install
@@ -165,6 +174,10 @@ install(
   DESTINATION ${INSTALL_INCLUDEDIR}
   FILES_MATCHING
   PATTERN "*.hpp"
+)
+
+install(FILES ${CMAKE_CURRENT_BINARY_DIR}/fft/src/KokkosFFT_config.h ${CMAKE_CURRENT_BINARY_DIR}/fft/src/KokkosFFT_Version_Info.hpp
+        DESTINATION ${INSTALL_INCLUDEDIR}
 )
 
 install(FILES ${CMAKE_CURRENT_SOURCE_DIR}/cmake/FindFFTW.cmake DESTINATION ${INSTALL_LIBDIR})
