+++ fft/src/CMakeLists.txt
@@ -10,42 +10,36 @@ target_link_libraries(fft INTERFACE common Kokkos::kokkos)
 if("CUFFT" IN_LIST KOKKOSFFT_TPL_LIST)
   find_package(CUDAToolkit REQUIRED COMPONENTS cufft)
   target_link_libraries(fft INTERFACE CUDA::cufft)
-  target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_CUFFT)
 endif()
 
 if("ROCFFT" IN_LIST KOKKOSFFT_TPL_LIST)
   find_package(rocfft REQUIRED)
   target_link_libraries(fft INTERFACE roc::rocfft)
-  target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_ROCFFT)
 endif()
 
 if("HIPFFT" IN_LIST KOKKOSFFT_TPL_LIST)
   find_package(hipfft REQUIRED)
   target_link_libraries(fft INTERFACE hip::hipfft)
-  target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_HIPFFT)
 endif()
 
 if("ONEMKL" IN_LIST KOKKOSFFT_TPL_LIST)
   find_package(MKL REQUIRED COMPONENTS SYCL)
   target_link_libraries(fft INTERFACE MKL::MKL_SYCL)
-  target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_ONEMKL)
 endif()
 
 if("FFTW_OPENMP" IN_LIST KOKKOSFFT_TPL_LIST)
   find_package(FFTW MODULE REQUIRED COMPONENTS FLOAT_LIB DOUBLE_LIB FLOAT_OPENMP_LIB DOUBLE_OPENMP_LIB)
   target_link_libraries(fft INTERFACE FFTW::Float FFTW::Double FFTW::FloatOpenMP FFTW::DoubleOpenMP)
-  target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_FFTW)
 elseif("FFTW_THREADS" IN_LIST KOKKOSFFT_TPL_LIST)
   find_package(FFTW MODULE REQUIRED COMPONENTS FLOAT_LIB DOUBLE_LIB FLOAT_THREADS_LIB DOUBLE_THREADS_LIB)
   target_link_libraries(fft INTERFACE FFTW::Float FFTW::Double FFTW::FloatThreads FFTW::DoubleThreads)
-  target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_FFTW)
 elseif("FFTW_SERIAL" IN_LIST KOKKOSFFT_TPL_LIST)
   find_package(FFTW MODULE REQUIRED COMPONENTS FLOAT_LIB DOUBLE_LIB)
   target_link_libraries(fft INTERFACE FFTW::Float FFTW::Double)
-  target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_FFTW)
 endif()
 
+message(STATUS "CMAKE_CURRENT_BINARY_DIR: ${CMAKE_CURRENT_BINARY_DIR}")
 target_include_directories(
-  fft INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}> $<INSTALL_INTERFACE:${INSTALL_INCLUDEDIR}>
[…]
