### 2025-06-27 aaf18bc
rename config.h to config.hpp based on review
