+++ fft/src/CMakeLists.txt
@@ -38,7 +38,6 @@ elseif("FFTW_SERIAL" IN_LIST KOKKOSFFT_TPL_LIST)
   target_link_libraries(fft INTERFACE FFTW::Float FFTW::Double)
 endif()
 
-message(STATUS "CMAKE_CURRENT_BINARY_DIR: ${CMAKE_CURRENT_BINARY_DIR}")
 target_include_directories(
   fft INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}> $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}> $<INSTALL_INTERFACE:${INSTALL_INCLUDEDIR}>
 )
