### 2025-06-27 de270be
remove unused macro
