### 2025-06-28 899404b
Improve error message of a failing find_package for FFTW
