+++ fft/src/CMakeLists.txt
@@ -32,15 +32,15 @@ if("ONEMKL" IN_LIST KOKKOSFFT_TPL_LIST)
 endif()
 
 if("FFTW_OPENMP" IN_LIST KOKKOSFFT_TPL_LIST)
-  find_package(FFTW MODULE REQUIRED)
+  find_package(FFTW MODULE REQUIRED COMPONENTS FLOAT_LIB DOUBLE_LIB FLOAT_OPENMP_LIB DOUBLE_OPENMP_LIB)
   target_link_libraries(fft INTERFACE FFTW::Float FFTW::Double FFTW::FloatOpenMP FFTW::DoubleOpenMP)
   target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_FFTW)
 elseif("FFTW_THREADS" IN_LIST KOKKOSFFT_TPL_LIST)
-  find_package(FFTW MODULE REQUIRED)
+  find_package(FFTW MODULE REQUIRED COMPONENTS FLOAT_LIB DOUBLE_LIB FLOAT_THREADS_LIB DOUBLE_THREADS_LIB)
   target_link_libraries(fft INTERFACE FFTW::Float FFTW::Double FFTW::FloatThreads FFTW::DoubleThreads)
   target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_FFTW)
 elseif("FFTW_SERIAL" IN_LIST KOKKOSFFT_TPL_LIST)
-  find_package(FFTW MODULE REQUIRED)
+  find_package(FFTW MODULE REQUIRED COMPONENTS FLOAT_LIB DOUBLE_LIB)
   target_link_libraries(fft INTERFACE FFTW::Float FFTW::Double)
   target_compile_definitions(fft INTERFACE KOKKOSFFT_ENABLE_TPL_FFTW)
 endif()
