### 2025-06-28 01e386d
KOKKOSFFT_VERSION should be defined from PROJECT_VERSION
