+++ CMakeLists.txt
@@ -22,10 +22,10 @@ option(KokkosFFT_ENABLE_DOCS "Build KokkosFFT documentaion/website" OFF)
 option(KokkosFFT_ENABLE_TESTING_TOOLS "Enable unit-testing tools" OFF)
 
 # Version information
-set(KOKKOSFFT_VERSION ${CMAKE_PROJECT_VERSION})
-set(KOKKOSFFT_VERSION_MAJOR ${CMAKE_PROJECT_VERSION_MAJOR})
-set(KOKKOSFFT_VERSION_MINOR ${CMAKE_PROJECT_VERSION_MINOR})
-set(KOKKOSFFT_VERSION_PATCH ${CMAKE_PROJECT_VERSION_PATCH})
+set(KOKKOSFFT_VERSION ${PROJECT_VERSION})
+set(KOKKOSFFT_VERSION_MAJOR ${PROJECT_VERSION_MAJOR})
+set(KOKKOSFFT_VERSION_MINOR ${PROJECT_VERSION_MINOR})
+set(KOKKOSFFT_VERSION_PATCH ${PROJECT_VERSION_PATCH})
 
 set(KOKKOS_REQUIRED_VERSION 4.5.0)
 
