### 2025-06-28 f797680
Update to kokkos 4.5.1 (#292)
* update kokkos to 4.5.1

* update kokkos to 4.5.01

* reduce parallelization in build

---------

Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
