+++ docs/intro/building.rst
@@ -130,7 +130,7 @@ if one of the device backend is enabled.
 Kokkos backends
 ---------------
 
-kokkos-fft requires ``Kokkos 4.4+``. For the moment, we support following backends for CPUs and GPUs.
+kokkos-fft requires ``Kokkos 4.5+``. For the moment, we support following backends for CPUs and GPUs.
 A FFT library dedicated to Kokkos Device backend (e.g. cufft for CUDA backend) is automatically used. 
 If CMake fails to find a backend FFT library, see :doc:`How to find fft libraries?<../finding_libraries>`.
 We may support experimental backends like ``OPENMPTARGET`` in the future.
