+++ CMakeLists.txt
@@ -26,7 +26,7 @@ set(KOKKOSFFT_VERSION_MAJOR ${PROJECT_VERSION_MAJOR})
 set(KOKKOSFFT_VERSION_MINOR ${PROJECT_VERSION_MINOR})
 set(KOKKOSFFT_VERSION_PATCH ${PROJECT_VERSION_PATCH})
 
-set(KOKKOS_REQUIRED_VERSION 4.4.0)
+set(KOKKOS_REQUIRED_VERSION 4.5.0)
 
 if(NOT KokkosFFT_ENABLE_INTERNAL_KOKKOS)
   # First check, Kokkos is added as subdirectory or not
