### 2025-06-29 4a0fd17
Include config.h in KokkosFFT_default_types.hpp (#291)
* Include config.h in KokkosFFT_default_types.hpp

* rename config.h to config.hpp based on review

* fix file name

* remove unused macro

* KOKKOSFFT_VERSION should be defined from PROJECT_VERSION
