+++ cmake/KokkosFFT_config.hpp.in
@@ -2,22 +2,14 @@
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0 WITH LLVM-exception
 
-#ifndef KOKKOSFFT_CONFIG_H
-#define KOKKOSFFT_CONFIG_H
+#ifndef KOKKOSFFT_CONFIG_HPP
+#define KOKKOSFFT_CONFIG_HPP
 
 /* Define the current version of Kokkos Kernels */
 #define KOKKOSFFT_VERSION @KOKKOSFFT_VERSION@
 #define KOKKOSFFT_VERSION_MAJOR @KOKKOSFFT_VERSION_MAJOR@
 #define KOKKOSFFT_VERSION_MINOR @KOKKOSFFT_VERSION_MINOR@
 #define KOKKOSFFT_VERSION_PATCH @KOKKOSFFT_VERSION_PATCH@
-
-/* Define if building in debug mode */
-#cmakedefine HAVE_KOKKOSFFT_DEBUG
-
-#cmakedefine KOKKOSFFT_ENABLE_BENCHMARK
-
-/* Define this macro if experimental features of KokkosFFT are enabled */
-#cmakedefine HAVE_KOKKOSFFT_EXPERIMENTAL
 
 /* FFTW */
 #cmakedefine KOKKOSFFT_ENABLE_TPL_FFTW
