### 2025-06-29 fd27f26
Allow index_sequence to work on unsigned integers (#287)
* Allow index_sequence to work on unsigned integers

* rename values in test_index_sequence test

* remove unused views from index_sequence test

---------

Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
