+++ common/unit_test/Test_Common_Utils.cpp
@@ -10,7 +10,8 @@ using execution_space = Kokkos::DefaultExecutionSpace;
 using test_types = ::testing::Types<Kokkos::LayoutLeft, Kokkos::LayoutRight>;
 
 // Int like types
-using base_int_types = ::testing::Types<int, std::size_t>;
+using base_int_types   = ::testing::Types<int, std::size_t>;
+using signed_int_types = ::testing::Types<int, std::int32_t, std::int64_t>;
 
 // value type combinations that are tested
 using paired_scalar_types = ::testing::Types<
@@ -21,8 +22,8 @@ using paired_scalar_types = ::testing::Types<
 
 // Basically the same fixtures, used for labeling tests
 template <typename T>
-struct ConvertNegativeAxis : public ::testing::Test {
-  using layout_type = T;
+struct TestConvertNegativeAxis : public ::testing::Test {
+  using value_type = T;
 };
 
 template <typename T>
@@ -50,45 +51,42 @@ struct TestIndexSequence : public ::testing::Test {
 };
 
 // Tests for convert_negative_axes over ND views
-template <typename LayoutType>
+template <typename IntType>
 void test_convert_negative_axes_1d() {
-  const int len        = 30;
-  using RealView1Dtype = Kokkos::View<double*, LayoutType, execution_space>;
-  RealView1Dtype x("x", len);
+  const IntType DIM = 1;
+  IntType converted_axis_0 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, IntType(0));
+  IntType converted_axis_minus1 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, IntType(-1));
 
-  int converted_axis_0 = KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/0);
-  int converted_axis_minus1 =
-      KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/-1);
-
-  int ref_converted_axis_0      = 0;
-  int ref_converted_axis_minus1 = 0;
+  IntType ref_converted_axis_0      = 0;
+  IntType ref_converted_axis_minus1 = 0;
 
   EXPECT_EQ(converted_axis_0, ref_converted_axis_0);
   EXPECT_EQ(converted_axis_minus1, ref_converted_axis_minus1);
 
   // Check if errors are correctly raised against invalid axis
   // axis must be in [-1, 1)
-  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/1); },
+  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, IntType(1)); },
                std::runtime_error);
 
-  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(x, /*axis=*/-2); },
+  EXPECT_THROW({ KokkosFFT::Impl::convert_negative_axis(DIM, IntType(-2)); },
                std::runtime_error);
 }
 
-template <typename LayoutType>
+template <typename IntType>
 void test_convert_negative_axes_2d() {
-  const int n0 = 3, n1 = 5;
-  using RealView2Dtype = Kokkos::View<double**, LayoutType, execution_space>;
-  RealView2Dtype x("x", n0, n1);
+  const IntType DIM = 2;
+  IntType converted_axis_0 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, IntType(0));
+  IntType converted_axis_1 =
+      KokkosFFT::Impl::convert_negative_axis(DIM, IntType(1));
[…]
