### 2025-06-30 8d80c02
hotfix: make convert_negative_axis a templated function again (#288)
* fix: conflict

* fix: conflict

* fix rank type

* fix: conflicts

---------

Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
