+++ .github/workflows/build_test.yaml
@@ -35,7 +35,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
-    - uses: crate-ci/typos@b1ae8d918b6e85bd611117d3d9a3be4f903ee5e4  # v1.33.1
+    - uses: crate-ci/typos@392b78fe18a52790c53f42456e46124f77346842  # v1.34.0
       with:
         files: ./cmake/ ./CMakeLists.txt ./docs/ ./README.md ./common/ ./fft/ ./examples/ ./install_test/ ./testing/
         config: ./.typos.toml
