### 2025-07-03 d41163f
add-navier-stokes-example
