+++ examples/advanced/01_NavierStokes/01_navier_stokes.cpp
@@ -1,0 +1,764 @@
+#include <vector>
+#include <iostream>
+#include <memory>
+#include <Kokkos_Core.hpp>
+#include <Kokkos_Complex.hpp>
+#include <Kokkos_Random.hpp>
+#include <KokkosFFT.hpp>
+#include "math_utils.hpp"
+#include "io_utils.hpp"
+
+constexpr int DIM = 3;
+
+using execution_space      = Kokkos::DefaultExecutionSpace;
+using host_execution_space = Kokkos::DefaultHostExecutionSpace;
+template <typename T>
+using View1D = Kokkos::View<T*, Kokkos::LayoutRight, execution_space>;
+template <typename T>
+using View2D = Kokkos::View<T**, Kokkos::LayoutRight, execution_space>;
+template <typename T>
+using View3D = Kokkos::View<T***, Kokkos::LayoutRight, execution_space>;
+template <typename T>
+using View4D = Kokkos::View<T****, Kokkos::LayoutRight, execution_space>;
+
+using range3D_type = Kokkos::MDRangePolicy<
+    execution_space,
+    Kokkos::Rank<3, Kokkos::Iterate::Right, Kokkos::Iterate::Right>,
+    Kokkos::IndexType<int>>;
+
+// \brief A class to represent the grid used in the Navier-Stokes equation.
+struct Grid {
+  //! Grid in x direction (nx)
+  View1D<double> m_x;
+
+  //! Grid in y direction (ny))
+  View1D<double> m_y;
+
+  //! Grid in z direction (nz)
+  View1D<double> m_z;
+
+  //! Wavenumber in x direction (nkx * 2 + 1)
+  View1D<double> m_kx;
+
+  //! Wavenumber in y direction (nky * 2 + 1)
+  View1D<double> m_ky;
+
+  //! Wavenumber in z direction (nkz + 1)
+  View1D<double> m_kzh;
+
+  //! Wavenumber squared (nkx * 2 + 1, nky * 2 + 1, nkz + 1)
+  View3D<double> m_ksq;
+
+  //! Inverse of k^2 for pressure projection (nkx * 2 + 1, nky * 2 + 1, nkz + 1)
+  View3D<double> m_inv_ksq;
+
+  //! Mask for dealiasing (nkx * 2 + 1, nky * 2 + 1, nkz + 1)
+  View3D<double> m_alias_mask;
+
+  // \brief Constructor of a Grid class
+  //
+  // \param[in] nx Number of grid points in the x-direction.
+  // \param[in] ny Number of grid points in the y-direction.
+  // \param[in] nz Number of grid points in the z-direction.
+  // \param[in] lx Length of the domain in the x-direction.
+  // \param[in] ly Length of the domain in the y-direction.
+  // \param[in] lz Length of the domain in the z-direction.
+  Grid(int nx, int ny, int nz, double lx, double ly, double lz) {
+    // Grid and Wavenumbers
+    execution_space exec;
+    m_x = Math::linspace(exec, 0.0, lx * M_PI, nx, /*endpoint=*/false);
+    m_y = Math::linspace(exec, 0.0, ly * M_PI, ny, /*endpoint=*/false);
+    m_z = Math::linspace(exec, 0.0, lz * M_PI, nz, /*endpoint=*/false);
+
+    // Wavenumbers
+    // int nkx2 = nx * 2 + 1, nky2 = ny * 2 + 1, nkzh = nz + 1;
+    int nzh      = nz / 2 + 1;
+    m_kx         = View1D<double>("kx", nx);
+    m_ky         = View1D<double>("ky", ny);
+    m_kzh        = View1D<double>("kzh", nzh);
+    m_ksq        = View3D<double>("ksq", nx, ny, nzh);
[…]
