+++ examples/advanced/CMakeLists.txt
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: (C) The kokkos-fft development team, see COPYRIGHT.md file
+#
+# SPDX-License-Identifier: MIT OR Apache-2.0 WITH LLVM-exception
+
+add_subdirectory(01_NavierStokes)
