+++ examples/advanced/01_NavierStokes/01_navier_stokes.cpp
@@ -1,4 +1,7 @@
-#include <vector>
+// SPDX-FileCopyrightText: (C) The Kokkos-FFT development team, see COPYRIGHT.md file
+//
+// SPDX-License-Identifier: MIT OR Apache-2.0 WITH LLVM-exception
+
 #include <iostream>
 #include <memory>
 #include <Kokkos_Core.hpp>
