### 2025-07-08 kokkos/kokkos-fft#297 @thivinanandh
Joss-Review Issue
## Review Summary

This library is cleanly written, well-structured, and comes with solid documentation. I was able to install it (as a subdirectory) without issues and could reproduce the performance comparisons mentioned. The continuous integration setup is particularly well done — I actually learned a few new tricks for handling CI in C++ projects.

A few small suggestions that might help improve the paper and the repo further:

### Statement of Need
It would be helpful to briefly mention other existing libraries that serve a similar purpose (e.g., FluidFFT) and highlight how this library differs — whether in performance, design choices, or usability. If this is the first of its kind library, then kindly mention the same in paper.

### Licensing
The README mentions just two licenses — MIT and Apache-2.0 with LLVM exception — but the `licenses/` folder includes a few others (BSD-3-Clause, CC0-1.0, etc.). It might be worth double-checking whether these are all relevant or if some can be cleaned up for clarity.

### Minor Enhancement
I submitted a small PR [#296](https://github.com/kokkos/kokkos-fft/pull/296) that adds the `sphinx-copybutton` extension to the docs, making it easier to copy code blocks. Please feel free to ignore it if you find it too minor for a PR — perhaps something to consider for a future release instead.
