### 2025-07-08 #8391 @thivinanandh
@yasahi-hpc:

From my review, I found kokkos-fft to be well written and documented library with excellent CI routines. I have finished my review and added my minor suggestions to the paper on this [issue](https://github.com/kokkos/kokkos-fft/issues/297).

Kindly go through the changes and address them if you found them to be usefull.

Please let me know if you need any additional information from myside.
