### 2025-07-09 50bcb20
"Docs: Added Sphinx_copybutton (#296)
- Added Sphinx copy button extension to enable easy copy of commands and code from documentation
