### 2025-07-09 #8391 @yasahi-hpc
> [@yasahi-hpc](https://github.com/yasahi-hpc):
>
> From my review, I found kokkos-fft to be well written and documented library with excellent CI routines. I have finished my review and added my minor suggestions to the paper on this [issue](https://github.com/kokkos/kokkos-fft/issues/297).
>
> Kindly go through the changes and address them if you found them to be usefull.
>
> Please let me know if you need any additional information from myside.

@thivinanandh

Thanks a lot for your review and making an issue.
Let me double check the license again.

I will update the Statement of Need section to mention similar libraries.
