### 2025-07-09 c587c05
Add dependency to sphinx-copybutton (#298)
Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
