+++ docs/requirements.txt
@@ -4,3 +4,4 @@
 
 breathe
 sphinx-rtd-theme
+sphinx-copybutton
