### 2025-07-09 20a4f1c
detail licenses in README.md
