### 2025-07-10 e5aa357
Detail license in README (#299)
* fix: typo in COPYRIGHT.md

* detail licenses in README.md

---------

Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
