+++ COPYRIGHT.md
@@ -4,12 +4,12 @@ SPDX-FileCopyrightText: (C) The kokkos-fft development team, see below
 SPDX-License-Identifier: MIT OR Apache-2.0 WITH LLVM-exception
 -->
 
-The kokkos-fft project licencing information follows the REUSE specification.
+The kokkos-fft project licensing information follows the REUSE specification.
 
 It is distributed under either:
 * the MIT License whose full text is available in the LICENSES/MIT.txt file,
 or, at your option,
-* the Apache-2.0 licence with LLVM exception whose full text is available in
+* the Apache-2.0 license with LLVM exception whose full text is available in
   the LICENSES/Apache-2.0.txt and LICENSES/LLVM-exception.txt files.
 
 For the purpose of copyright, the kokkos-fft development team is defined as follow:
