### 2025-07-10 ece7069
Add installation guide with spack
