+++ paper/paper.bib
@@ -60,6 +60,25 @@ month=sep}
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
+@article{Mohanan2019,
+ abstract = {The Python package fluidfft provides a common Python API for performing Fast Fourier Transforms (FFT) in sequential, in parallel and on GPU with different FFT libraries (FFTW, P3DFFT, PFFT, cuFFT). fluidfft is a comprehensive FFT framework which allows Python users to easily and efficiently perform FFT and the associated tasks, such as computing linear operators and energy spectra. We describe the architecture of the package composed of C++ and Cython FFT classes, Python “operator” classes and Pythran functions. The package supplies utilities to easily test itself and benchmark the different FFT solutions for a particular case and on a particular machine. We present a performance scaling analysis on three different computing clusters and a microbenchmark showing that fluidfft is an interesting solution to write efficient Python applications using FFT. Funding statement: This project has indirectly benefited from funding from the foundation Simone et Cino Del Duca de l’Institut de France, the European Research Council (ERC) under the European Union’s Horizon 2020 research and innovation program (grant agreement No 647018-WATU and Euhit consortium) and the Swedish Research Council (Vetenskapsrådet): 2013–5191. We have also been able to use supercomputers of CIMENT/GRICAD, CINES/GENCI (grant 2018-A0040107567) and the Swedish National Infrastructure for Computing (SNIC).},
+ author = {Mohanan, Ashwin Vishnu and Bonamy, Cyrille and Augier, Pierre},
+ doi = {10.5334/jors.238},
+ journal = {Journal of Open Research Software},
+ keyword = {en_US},
+ month = {Apr},
+ title = {FluidFFT: Common API (C++ and Python) for Fast Fourier Transform HPC Libraries},
+ year = {2019}
+}
+
+@software{jax2018github,
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
+  url = {http://github.com/jax-ml/jax},
+  version = {0.3.13},
+  year = {2018},
+}
+
 @article{Wakatani1984,
     author = {Wakatani, Masahiro and Hasegawa, Akira},
     title = {A collisional drift wave description of plasma edge turbulence},
