### 2025-07-11 af00a84
Introduce spack installation test
