+++ .github/workflows/spack-test.yaml
@@ -38,8 +38,8 @@ jobs:
       - name: Configure and build with CMake
         run: |
           docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
-          source /opt/spack/share/spack/setup-env.sh && \
-          spack install kokkos ${{ matrix.backend.kokkos }} && \
-          spack install kokkos-fft ${{ matrix.backend.kokkos_fft }} && \
-          cmake -B install_test/as_library/build && \
-          cmake --build install_test/as_library/build -j 4
+          bash -c "source /opt/spack/share/spack/setup-env.sh && \
+            spack install kokkos ${{ matrix.backend.kokkos }} && \
+            spack install kokkos-fft ${{ matrix.backend.kokkos_fft }} && \
+            cmake -B install_test/as_library/build && \
+            cmake --build install_test/as_library/build -j 4"
