### 2025-07-11 34780d6
fix spack workflow
