+++ .github/workflows/spack-test.yaml
@@ -33,16 +33,16 @@ jobs:
       - name: Clone and install spack
         run: |
           docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
-          git clone --depth=2 https://github.com/spack/spack.git /opt/spack
+          git clone --depth=2 https://github.com/spack/spack.git
           docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
-          ls /opt/spack/
+          ls spack/
           docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
-          /opt/spack/bin/spack install libelf
+          spack/bin/spack install libelf
 
       - name: Configure and build with CMake
         run: |
           docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
-          bash -c "source /opt/spack/share/spack/setup-env.sh && \
+          bash -c "source spack/share/spack/setup-env.sh && \
             spack install kokkos ${{ matrix.backend.kokkos }} && \
             spack install kokkos-fft ${{ matrix.backend.kokkos_fft }} && \
             cmake -B install_test/as_library/build && \
