+++ .github/workflows/spack-test.yaml
@@ -34,10 +34,4 @@ jobs:
           bash -c "git clone --depth=2 --branch=v1.0.1 https://github.com/spack/spack.git"
 
       - name: Configure and build with CMake
-        run: |
           docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
-          bash -c "source spack/share/spack/setup-env.sh && \
-            spack install kokkos-fft +tests ^kokkos ${{ matrix.backend.kokkos }} && \
-            spack load kokkos kokkos-fft && \
-            cmake -B install_test/as_library/build && \
-            cmake --build install_test/as_library/build -j 4"
