+++ .github/workflows/spack-test.yaml
@@ -33,10 +33,12 @@ jobs:
       - name: Checkout built branch
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
 
-      - name: Set up Python
-        uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065  # v5.6.0
-        with:
-          python-version: 3.12
+      - name: Install Python
+        run: |
+          docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
+          apt-get update && apt-get install -y \
+          python3.12 \
+          && rm -rf /var/lib/apt/lists/*
 
       - name: Clone and install spack
         run: |
