### 2025-07-15 51e2133
just rely on spack
