+++ .github/workflows/spack-test.yaml
@@ -15,45 +15,46 @@ jobs:
       matrix:
         backend:
           - name: nvcc
-            image: nvcc
             kokkos: +cuda +wrapper cuda_arch=90
             kokkos_fft: cufft
           - name: rocm
-            image: rocm
             kokkos: +rocm amdgpu_target=gfx90a
             kokkos_fft: hipfft
 
     steps:
-      - name: Free Disk Space (Ubuntu)
-        uses: jlumbroso/free-disk-space@54081f138730dfa15788a46383842cd2f914a1be  # v1.3.1
-        with:
-          tool-cache: true
-          large-packages: false
-
       - name: Checkout built branch
         uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683  # v4.2.2
 
-      - name: Install Python
-        run: |
-          docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
-          apt-get update && apt-get install -y \
-          python3 \
-          python3-pip \
-          python3-venv \
-          && rm -rf /var/lib/apt/lists/*
+      - name: Set up Python
+        uses: actions/setup-python@a26af69be951a213d495a4c3e4e4022e16d87065  # v5.6.0
+        with:
+          python-version: 3.12
 
       - name: Clone and install spack
         run: |
-          docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
           git clone --depth=2 https://github.com/spack/spack.git
-          docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
           spack/bin/spack install libelf
+
+      - name: Install cmake
+        run: |
+          bash -c "source spack/share/spack/setup-env.sh && \
+            spack install cmake
+
+      - name: Install kokkos
+        run: |
+          bash -c "source spack/share/spack/setup-env.sh && \
+            spack load cmake && \
+            spack install kokkos ${{ matrix.backend.kokkos }}
+
+      - name: Install kokkos-fft
+        run: |
+          bash -c "source spack/share/spack/setup-env.sh && \
+            spack load cmake && spack load kokkos \
+            spack install kokkos-fft ${{ matrix.backend.kokkos_fft }}
 
       - name: Configure and build with CMake
         run: |
-          docker run -v ${{ github.workspace }}:/work ghcr.io/kokkos/kokkos-fft/base_${{ matrix.backend.image }}:latest \
           bash -c "source spack/share/spack/setup-env.sh && \
-            spack install kokkos ${{ matrix.backend.kokkos }} && \
-            spack install kokkos-fft ${{ matrix.backend.kokkos_fft }} && \
+            spack load cmake && spack load kokkos && spack load kokkos-fft \
             cmake -B install_test/as_library/build && \
             cmake --build install_test/as_library/build -j 4"
