### 2025-07-15 41cddba
fix spack setup
