+++ .github/workflows/spack-test.yaml
@@ -37,24 +37,24 @@ jobs:
 
       - name: Install cmake
         run: |
-          bash -c "source spack/share/spack/setup-env.sh && \
-            spack install cmake
+          . ./spack/share/spack/setup-env.sh
+          spack install cmake
 
       - name: Install kokkos
         run: |
-          bash -c "source spack/share/spack/setup-env.sh && \
-            spack load cmake && \
-            spack install kokkos ${{ matrix.backend.kokkos }}
+          . ./spack/share/spack/setup-env.sh
+          spack load cmake
+          spack install kokkos ${{ matrix.backend.kokkos }}
 
       - name: Install kokkos-fft
         run: |
-          bash -c "source spack/share/spack/setup-env.sh && \
-            spack load cmake && spack load kokkos \
-            spack install kokkos-fft ${{ matrix.backend.kokkos_fft }}
+          . ./spack/share/spack/setup-env.sh
+          spack load cmake kokkos
+          spack install kokkos-fft ${{ matrix.backend.kokkos_fft }}
 
       - name: Configure and build with CMake
         run: |
-          bash -c "source spack/share/spack/setup-env.sh && \
-            spack load cmake && spack load kokkos && spack load kokkos-fft \
-            cmake -B install_test/as_library/build && \
-            cmake --build install_test/as_library/build -j 4"
+          . ./spack/share/spack/setup-env.sh
+          spack load cmake kokkos kokkos-fft
+          cmake -B install_test/as_library/build && \
+          cmake --build install_test/as_library/build -j 4"
