### 2025-07-22 c8f793a
Print in/out view details and axes in case of extent errors
