### 2025-07-22 8bd6f87
capture views, axes and rank explicitly
