### 2025-07-22 fbc4bdc
Do not capture rank
