### 2025-07-22 dc41a73
Print in/out view details and axes in case of extent errors (#301)
* Print in/out view details and axes in case of extent errors

* capture views, axes and rank explicitly

* Do not capture rank

---------

Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
