### 2025-07-28 #8391 @danielskatz
Thanks @jpmorgan98. We'll go ahead without you then.
