### 2025-07-28 41cdff9
small changes in paper.md
