+++ paper/paper.md
@@ -49,29 +49,29 @@ kokkos-fft provides a unified, performance-portable interface for Fast Fourier T
 
 The fast Fourier transform (FFT) is a family of fundamental algorithms that is widely used in scientific computing and other areas [@Rockmore2000]. [kokkos-fft](https://github.com/kokkos/kokkos-fft) is designed to help [Kokkos](https://github.com/kokkos/kokkos) [@Trott2022] users who are:
 
-* developing a Kokkos application which relies on FFT libraries. E.g., fluid simulation codes with periodic boundaries, plasma turbulence, etc.
+* developing a Kokkos application that relies on FFT libraries, e.g., fluid simulation codes with periodic boundaries, plasma turbulence, etc.
 
-* inclined to integrate in-situ signal and image processing with FFTs. E.g., spectral analyses, low pass filtering, etc.
+* inclined to integrate in-situ signal and image processing with FFTs, e.g., spectral analyses, low pass filtering, etc.
 
 * willing to use de facto standard FFT libraries just like [`numpy.fft`](https://numpy.org/doc/stable/reference/routines.fft.html) [@Harris2020].
 
 kokkos-fft can benefit such users through the following features:
 
 * A simple interface like [`numpy.fft`](https://numpy.org/doc/stable/reference/routines.fft.html) with in-place and out-of-place transforms:  
-Only accepts [Kokkos Views](https://kokkos.org/kokkos-core-wiki/API/core/view/view.html) which corresponds to the [numpy.array](https://numpy.org/doc/stable/reference/generated/numpy.array.html), to make APIs simple and safe.
+Only accepts [Kokkos Views](https://kokkos.org/kokkos-core-wiki/API/core/view/view.html) that correspond to the [numpy.array](https://numpy.org/doc/stable/reference/generated/numpy.array.html), to make APIs simple and safe.
 
 * 1D, 2D, 3D standard and real FFT functions (similar to [`numpy.fft`](https://numpy.org/doc/stable/reference/routines.fft.html)) over 1D to 8D Kokkos Views:  
 Batched plans are automatically used if View dimension is larger than FFT dimension.
 
-* A reusable [FFT plan](https://kokkosfft.readthedocs.io/en/latest/api/plan/plan.html) which wraps the vendor libraries for each Kokkos backend:  
+* A reusable [FFT plan](https://kokkosfft.readthedocs.io/en/latest/api/plan/plan.html) that wraps the vendor libraries for each Kokkos backend:  
 [FFTW](http://www.fftw.org), [cuFFT](https://developer.nvidia.com/cufft), [rocFFT](https://github.com/ROCm/rocFFT), and [oneMKL](https://www.intel.com/content/www/us/en/developer/tools/oneapi/onemkl.html) are automatically enabled based on the enabled Kokkos backend.
 
 * Support for multiple CPU and GPU backends:  
-FFT libraries for the enabled Kokkos backend are executed on the stream/queue used in that [`ExecutionSpace`](https://kokkos.org/kokkos-core-wiki/API/core/execution_spaces.html) where the parallel operations are performed.
+FFT libraries for the enabled Kokkos backend are executed on the stream/queue used in the [`ExecutionSpace`](https://kokkos.org/kokkos-core-wiki/API/core/execution_spaces.html) where the parallel operations are performed.
 
-* Compile time and/or runtime errors for invalid usage (e.g. `View` extents mismatch).
+* Compile time and/or runtime errors for invalid usage (e.g., `View` extents mismatch).
 
[…]
