### 2025-07-28 844ca02
minor changes in paper.bib
