+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Trott2021,
 author={Trott, Christian and Berger-Vergiat, Luc and Poliakoff, David and Rajamanickam, Sivasankaran and Lebrun-Grandie, Damien and Madsen, Jonathan and Al Awar, Nader and Gligoric, Milos and Shipman, Galen and Womeldorff, Geoff},
 journal={ Computing in Science \& Engineering },
-title={{ The Kokkos EcoSystem: Comprehensive Performance Portability for High Performance Computing }},
+title={The {K}okkos EcoSystem: Comprehensive Performance Portability for High Performance Computing},
 year={2021},
 volume={23},
 number={05},
@@ -67,7 +67,7 @@ month=sep}
  journal = {Journal of Open Research Software},
  keyword = {en_US},
  month = {Apr},
- title = {FluidFFT: Common API (C++ and Python) for Fast Fourier Transform HPC Libraries},
+ title = {FluidFFT: Common API ({C}++ and {P}ython) for {F}ast {F}ourier {T}ransform HPC Libraries},
  year = {2019}
 }
 
