### 2025-07-28 #8391 @danielskatz
@yasahi-hpc - I'm suggesting some minor changes in the paper in https://github.com/kokkos/kokkos-fft/pull/302  Please merge this, or let me know what you disagree with, then we can continue the process
