### 2025-07-29 #8391 @yasahi-hpc
> [@yasahi-hpc](https://github.com/yasahi-hpc) - I'm suggesting some minor changes in the paper in [kokkos/kokkos-fft#302](https://github.com/kokkos/kokkos-fft/pull/302) Please merge this, or let me know what you disagree with, then we can continue the process

Thank you for the suggestions. I have merged your PR
