### 2025-07-29 6be5ff9
Update version in CMake and CHANGELOG.md
