### 2025-07-30 f3c63ae
fix typo in CHANGELOG.md
