### 2025-07-30 30c535e
Move #288 from Bug fix to general enhancements in CHANGELOG.md
