### 2025-07-30 fa8fc05
Release candidate 0.4.0 (#303)
* Update version in CMake and CHANGELOG.md

* fix typo in CHANGELOG.md

* Move #288 from Bug fix to general enhancements in CHANGELOG.md

---------

Co-authored-by: Yuuichi Asahi <y.asahi@nr.titech.ac.jp>
