### 2025-07-30 #8391 @danielskatz
Congratulations to @yasahi-hpc (Yuuichi ASAHI) and co-authors on your JOSS publication!!

And thanks to @thivinanandh and @VasanthRajendran for reviewing!
